moved addable and setting elements to floating widget folder
