+++ .gitignore
@@ -2,6 +2,7 @@
 
 .npmrc
 *~
+/.idea
 
 # dependencies
 /node_modules
