+++ src/pages/AddCommunityPage/components/FloatingWidget/setting-elements.js
